equipments list

+++ src/app/entity-list/entity-list.component.ts
@@ -1,5 +1,6 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
 import { EquipmentService } from '../shared/equipments.service';
+import { Entity } from '../shared/models';
 
 @Component({
   selector: 'app-entity-list',
@@ -8,14 +9,29 @@ import { EquipmentService } from '../shared/equipments.service';
 })
 export class EntityListComponent implements OnInit {
 
+  @Input() source: string;
+
   constructor(public service: EquipmentService) { }
+
+  get entity(): Entity[] {
+    if (this.source === 'components') {
+      return this.service.selectedEquipment.components;
+    }
+    if (this.source === 'equipments') {
+      return this.service.equipments;
+    }
+  }
 
   ngOnInit(): void {
   }
 
   delete(id: number) {
-    //this.service.deleteEquipment(id);
-    this.service.removeComponent(id);
+    if (this.source === 'components') {
+      this.service.removeComponent(id);
+    }
+    if (this.source === 'equipments') {
+      this.service.removeEquipment(id);
+    }
   }
 
   save() {
